requirements.txt
python-dotenv
pandas
numpy
google-api-python-client
google-auth
google-auth-oauthlib
detoxify
streamlit>=1.28.0
plotly
tqdm
requests